requirements.txt
numpy
matplotlib
jupyterlab
PYYaml
pytest
pytest-cov